Extract pbi sensitivity label information
