+++ metaphor/common/dataclass.py
@@ -1,11 +1,8 @@
-from pydantic import Extra
-
-
 class ConnectorConfig:
     """
     Pydantic dataclass Config for metaphor connector configurations
     """
 
-    validate_all = True
-    extra = Extra.forbid
+    validate_default = True
+    extra = "forbid"
     coerce_numbers_to_str = True
